% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{indiv_harm_fits}
\alias{indiv_harm_fits}
\title{Importance of Individual Harmonics}
\usage{
indiv_harm_fits(theta, burn = 1000, thin = 1, alpha = 0.01)
}
\arguments{
\item{theta}{T x p x mc array of MCMC samples of theta}

\item{burn}{number of MCMC samples to use as burn-in}

\item{thin}{number of iterations to thin MCMC samples}

\item{alpha}{number specifying level of credible interval, e.g. alpha = 0.05 for 95\% credible interval}
}
\value{
List containing credible intervals for each harmonic state at each timepoint.
\describe{
\item{ints}{array containing credible interval bounds for each state.}
\item{props}{proportion of credible intervals which did not contain 0 for each harmonic.}
}
}
\description{
Function to compute credible intervals for each state to assess importance of individual harmonics
}
