% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_attributes}
\alias{location_attributes}
\title{Grid Cell Data}
\format{A data frame with 2163 rows and 4 columns:
\describe{
  \item{lon}{Longitude of the grid cell}
  \item{lat}{Latitude of the grid cell}
  \item{landcover}{Majority landcover classification for the grid cell, derived from the GLC 2000 dataset (\url{http://forobs.jrc.ec.europa.eu/products/glc2000/products.php})}
  \item{loc}{Index for the grid cell, matches column names of \code{mtci_data}}
}}
\usage{
location_attributes
}
\description{
A dataset containing latitude, longitude, landcover and cell index complementing the data in \code{mtci_data}.
}
\keyword{datasets}
