% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{MCMC_to_pheno_events}
\alias{MCMC_to_pheno_events}
\title{Phenological Event Estimation}
\usage{
MCMC_to_pheno_events(output, q = 2, burnin = 1000, thin = 2,
  cf = NULL, cf_calc = TRUE, s_cf = 2, shift = 8, d = 5,
  cut2 = 0.8, alpha = 0.1, s = 46, year_start = 2003,
  year_end = 2007, return_points = FALSE)
}
\arguments{
\item{output}{MCMC ouput from multiple chains, use \code{par_to_match()} around list of multiple chains}

\item{q}{number of harmonics included in the model}

\item{burnin}{number of iterations used for burnin}

\item{thin}{number of iterations to use to thin the chains (\code{thin = 1} keeps all samples)}

\item{cf}{value specifying the minimum difference between sos and pog in identification rule. If \code{cf = NULL} and \code{cf_calc=TRUE}, \code{cf} is s_cf*posterior mode of \eqn{\sigma_e}, else \code{cf} \eqn{\neq} NULL allows user to specify an alternate value for the cutoff.}

\item{cf_calc}{boolean where TRUE calculates \code{cf} as s_cf*posterior mode of \eqn{\sigma_e} and FALSE requires a specified cutoff value for \code{cf}.}

\item{s_cf}{value specifying the multiple of the posterior mode of \eqn{\sigma_e} to use in the computation of \code{cf} if \code{cf_calc = TRUE}}

\item{shift}{number of composites past January 1 to consider for the previous year (protects against POG that occur around Jan 1st).}

\item{d}{value specifying the minimum duration of a growing season}

\item{alpha}{value specifying the confidence level for (1 - alpha)*100\% intervals for phenological event timings}

\item{s}{period of harmonic time series}

\item{year_start}{value specifying the first year of the time range}

\item{year_end}{value specifying the last year of the time range}

\item{return_points}{(TRUE/FALSE) should events identified from all MCMC draws be returned, or just summaries?}
}
\value{
List containing:
\describe{
\item{intervals}{estimates and intervals for all identified phenological events}
\item{probs}{estimated probabilities of multiple growing seasons in each year}
\item{nseas_year}{highest probability number of growing seasons each year}
\item{nused}{number of MCMC samples used to estimate each phenological events}
\item{fits}{posterior median estimates and credible intervals of \eqn{SVI_{1:T}}}
\item{pts}{if \code{return_points = TRUE}, contains estimated events from all MCMC samples}
\item{var_summaries}{posterior summaries for observation and state equation standard deviations (\eqn{\sigma_e}, \eqn{\sigma_w})}
}
}
\description{
Function to estimate phenological events from posterior samples obtained from multiple MCMC chains
}
\examples{
data(mtci_data) # attach MTCI data

c.mtci <- scale(mtci_data, scale=FALSE) # center MTCI data

ps <- c(10,200) # initial starting precision values

y <- c.mtci[,5] # pull out 5th location

### Run multiple chains
twochains <- lapply(1:2, function(j) Rmcmc_DLM_half_Cauchy(y, sig2_e0=1/ps[j], sig2_w0=1/ps[j], 
                                                           s=46, seas=3, mc=1000, thin=1))  
# period s = 46
# include first 3 harmonics (seas = 3)
# run 1000 MCMC iterations (mc = 1000)
# no thinning, (thin = 1)
  
mcmc_smp <- par_to_match(twochains) # organize output from two chains

### Estimate phenological events
fit_sum = MCMC_to_pheno_events(mcmc_smp, shift=8, thin=2, d=8, s_cf=3, burnin=100, q=3)

}
