% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtci_data}
\alias{mtci_data}
\title{MTCI Data}
\format{A data frame with 230 rows and 2163 columns.}
\usage{
mtci_data
}
\description{
A dataset containing MTCI values for 2003 - 2008 on a grid over southern India. Each column corresponds to a grid cell in the region, each row corresponds to an 8-day composite measurement. Data were obtained from the NERC Earth Observation Data Centre (\url{www.neodc.rl.ac.uk/}).
}
\keyword{datasets}
