% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{par_to_match}
\alias{par_to_match}
\title{Combine MCMC Output from chains}
\usage{
par_to_match(mod_out)
}
\arguments{
\item{mod_out}{List of MCMC output from running multiple MCMC chains using \code{lapply()}, or \code{doParallel::mclapply()}}
}
\value{
List of organized MCMC samples.
\describe{
\item{sigma2_e}{mc x nchains matrix containing samples of \eqn{\sigma_e^2}}
\item{sigma2_w}{mc x nchains matrix containing samples of \eqn{\sigma_w^2}}
\item{theta}{list of length nchains containing samples of \eqn{\theta}}
}
}
\description{
Function to rearrange output from multiple MCMC chains.
}
\examples{
data(mtci_data) # attach MTCI data

c.mtci <- scale(mtci_data, scale=FALSE) # center MTCI data

ps <- c(10,200) # initial starting precision values

y <- c.mtci[,5] # pull out 5th location

### Run multiple chains
twochains <- lapply(1:2, function(j) Rmcmc_DLM_half_Cauchy(y, sig2_e0=1/ps[j], sig2_w0=1/ps[j], 
                                                           s=46, seas=3, mc=1000, thin=1))  
fits <- par_to_match(twochains) # organize output from two chains
}
