\name{phenoDLM-package}
\alias{phenoDLM-package}
\alias{phenoDLM}
\docType{package}
\title{
  Phenological event estimation using Bayesian DLMs.
}
\description{
This package provides functionality to implement the Bayesian DLM approach described in Johnson et al., (2019) for phenological event estimation from remotely sensed vegetation index (VI) data products.
}

\author{
Maggie Johnson

Maintainer: Maggie Johnson <maggie.johnson510@gmail.com>
}
\references{
Johnson, M.C., Caragea, P.C., Meiring, W., Jeganathan, C., Atkinson, P. (2019) Bayesian Dynamic Linear Models for Estimation of Phenological Events from Remote Sensing Data, \emph{JABES}, 24(1), 1-25. DOI: 10.1007/s13253-018-00338-y
}
