% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoDLM.R
\docType{package}
\name{phenoDLM}
\alias{phenoDLM}
\alias{phenoDLM-package}
\title{phenoDLM: A package for phenological event estimation using Bayesian DLMs.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{phenoDLM functions}{

The foo functions ...
}

