% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(y, fit)
}
\arguments{
\item{y}{\emph{T}-dim vector of observations.}

\item{fit}{object containing posterior summaries of \eqn{SVI_{1:T}} as obtained from \code{MCMC_to_pheno_events()}}
}
\value{
List containing:
\describe{
\item{rmse}{overall rmse}
\item{rmse_year}{rmse computed separately for each year}
}
}
\description{
Function to compute root mean squared error.
}
