% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{simulate_dlm_seas}
\alias{simulate_dlm_seas}
\title{Simulate Seasonal DLM}
\usage{
simulate_dlm_seas(T, theta_0, F, G, V, W)
}
\arguments{
\item{T}{number specifying the length of the simulated time series}

\item{theta_0}{p-dim vector of initial states}

\item{F}{observation equation matrix}

\item{G}{state equation transition matrix}

\item{V}{value specifying observation equation variance}

\item{W}{value specifying state equation variance}
}
\value{
List of simulated values:
\describe{
\item{y}{vector of simulated observations}
\item{v}{vector of simulated observation errors}
\item{w}{vector of simulated innovation errors}
\item{theta}{matrix of simulated state vectors}
}
}
\description{
Function to simulate time series from a univariate dynamic linear model with static \eqn{F}, \eqn{G}, \eqn{V}, \eqn{W}.
}
